+++ packages/package-template-nodejs/src/index.ts
@@ -1,0 +1,3 @@
+export function templateNodejs(): string {
+    return 'This package named `@easy/package-template-nodejs` is used for creating new nodejs packages.';
+}