+++ packages/project-template-browser-react/src/redux/thunks/createMainThunk.ts
@@ -1,0 +1,5 @@
+import { thunkCreatorFactory } from '@easy/utils-redux';
+
+import { IStoreState } from '../types';
+
+export const createMainThunk = thunkCreatorFactory<IStoreState>();