+++ packages/template-nodejs/src/index.ts
@@ -1,3 +1,0 @@
-export function templateNodejs(): string {
-    return 'This package named `@packages/template-nodejs` is used for creating new nodejs packages.';
-}