+++ projects/template-browser-react/src/redux/hooks/useMainSelector.ts
@@ -1,5 +1,0 @@
-import { createTypedSelector } from '@package/utils-redux';
-
-import { IStoreState } from '../types';
-
-export const useMainSelector = createTypedSelector<IStoreState>();