+++ projects/template-browser-react/src/redux/thunks/createMainThunk.ts
@@ -1,5 +1,0 @@
-import { thunkCreatorFactory } from '@package/utils-redux';
-
-import { IStoreState } from '../types';
-
-export const createMainThunk = thunkCreatorFactory<IStoreState>();