+++ projects/template-nodejs/src/global.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="@package/global-types" />