+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,5 @@
 # CHANGELOG
-## @packages/hooks v0.2.0 (Unreleased)
+## @easy/hooks v0.2.0 (Unreleased)
 ### Features
 
 - The parameter `initialValue` of `useBoolean` now is optional
@@ -8,7 +8,11 @@
 
 - Move tests into __tests__, rename testUtils.ts to utils.test.ts
 
-## @packages/hooks v0.1.0 (2021-09-30)
+### Breaking Changes
+
+- Renamed from `@package/hooks` to `@easy/hooks`
+
+## @package/hooks v0.1.0 (2021-09-30)
 ### Features
 
 - `useBoolean`