+++ packages/memorize/CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+## @easy/memorize v2.0.0 (Unreleased)
+### Breaking Changes
+
+- Renamed from `@package/memorize` to `@easy/memorize`
+
 ## @package/memorize v1.0.0 (2021-08-19)
 ### Features