+++ packages/random-string/CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+## @easy/random-string v1.0.0 (Unreleased)
+### Breaking Changes
+
+- Renamed from `@package/random-string` to `@easy/random-string`
+
 ## @package/random-string v1.0.0 (2021-08-24)
 ### Features