+++ packages/utils-fluentui/CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+## @easy/utils-fluentui v0.2.0 (Unreleased)
+### Breaking Changes
+
+- Renamed from `@package/utils-fluentui` to `@easy/utils-fluentui`
+
 ## @package/utils-fluentui v0.1.0 (2021-10-02)
 ### Features