+++ packages/utils-redux/CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+## @easy/utils-redux v1.0.0 (Unreleased)
+### Breaking Changes
+
+- Renamed from `@packages/utils-redux` to `@easy/utils-redux`
+
 ## @packages/utils-redux v1.0.0 (2021-09-30)
 ### Features